Unwind: replace pure virtual functions with aborts
Convert all pure virtual functions in the UnwindCursor with implementations that
abort.  This is effectively manually replicating the current behaviour, whilst
removing the compiler generated calls to __cxa_pure_virtual, which will abort at
runtime with a message indicating that a pure virtual call was made.

The whitespace changes are the result of executing clang-format over the changed
region.

git-svn-id: http://llvm.org/svn/llvm-project/libcxxabi/trunk@228423 91177308-0d34-0410-b5e6-96231b3b80d8